+++ gui/src/localization.ts
@@ -2,6 +2,7 @@ import i18n from 'i18next'
 import { initReactI18next } from 'react-i18next'
 import en from './locales/en/strings.json'
 import es from './locales/es/strings.json'
+import ru from './locales/ru/strings.json'
 
 i18n
   .use(initReactI18next) // passes i18n down to react-i18next
@@ -13,6 +14,9 @@ i18n
       es: {
         translation: es,
       },
+	  ru: {
+		translation: ru,
+	  },
     },
     lng: 'en',
     fallbackLng: 'en',